factors out the cloudhsm code into a pkg
creates a hsm.Client interface so that the cloudhsm specific code can be
potentially mocked out allowing us to test the controller specific code
without the need of spinning up an HSM in AWS
