+++ pkg/controller/controller.go
@@ -17,16 +17,19 @@ limitations under the License.
 package controller
 
 import (
+	"github.com/alphagov/verify-metadata-controller/pkg/hsm"
+	"github.com/alphagov/verify-metadata-controller/pkg/hsm/awscloudhsm"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 )
 
 // AddToManagerFuncs is a list of functions to add all Controllers to the Manager
-var AddToManagerFuncs []func(manager.Manager) error
+var AddToManagerFuncs []func(manager.Manager, hsm.Client) error
 
 // AddToManager adds all Controllers to the Manager
 func AddToManager(m manager.Manager) error {
+	hsmClient := &awscloudhsm.Client{}
 	for _, f := range AddToManagerFuncs {
-		if err := f(m); err != nil {
+		if err := f(m, hsmClient); err != nil {
 			return err
 		}
 	}
